Fixed: Vunerebility

+++ src/main/java/fact/it/edge_microservice/controller/PlayerController.java
@@ -44,14 +44,14 @@ public class PlayerController {
 
         List<PlayerData> playerDatas = responseEntityPlayerDatas.getBody();
 
-        assert playerDatas != null;
+        // assert playerDatas != null;
         for (PlayerData playerData: playerDatas) {
 
             TypeTamagotchi typeTamagotchi =
                     restTemplate.getForObject(URL_PROTOCOL + typeTamagotchiServiceBaseUrl + "/types/{typeName}",
                             TypeTamagotchi.class, playerData.getTypeName());
 
-            assert typeTamagotchi != null;
+            // assert typeTamagotchi != null;
             Player playerToAdd = new Player(playerData, typeTamagotchi);
 
             returnList.add(playerToAdd);
@@ -72,12 +72,12 @@ public class PlayerController {
 
         PlayerData playerData = responseEntityPlayerDatas.getBody();
 
-        assert playerData != null;
+        // assert playerData != null;
         TypeTamagotchi typeTamagotchi =
                 restTemplate.getForObject(URL_PROTOCOL + typeTamagotchiServiceBaseUrl + "/types/{typeName}",
                         TypeTamagotchi.class, playerData.getTypeName());
 
-        assert typeTamagotchi != null;
+        // assert typeTamagotchi != null;
         return new Player(playerData, typeTamagotchi);
     }
 
@@ -95,13 +95,13 @@ public class PlayerController {
 
         List<PlayerData> playerDatas = responseEntityPlayerDatas.getBody();
 
-        assert playerDatas != null;
+        // assert playerDatas != null;
         for (PlayerData playerData: playerDatas) {
             TypeTamagotchi typeTamagotchi =
                     restTemplate.getForObject(URL_PROTOCOL + typeTamagotchiServiceBaseUrl + "/types/{typeName}",
                             TypeTamagotchi.class, playerData.getTypeName());
 
-            assert typeTamagotchi != null;
+            // assert typeTamagotchi != null;
             returnList.add(new Player(playerData, typeTamagotchi));
         }
 
@@ -124,8 +124,8 @@ public class PlayerController {
                         TypeTamagotchi.class, typeName);
 
 
-        assert playerData != null;
-        assert typeTamagotchi != null;
+        // assert playerData != null;
+        // assert typeTamagotchi != null;
         return new Player(playerData, typeTamagotchi);
     }
 
@@ -147,7 +147,7 @@ public class PlayerController {
         PlayerData playerData =
                 restTemplate.getForObject(URL_PROTOCOL + playerDataServiceBaseUrl + "/playerData/" + playerDataCode,
                         PlayerData.class);
-        assert playerData != null;
+        // assert playerData != null;
         playerData.setName(name);
 
         ResponseEntity<PlayerData> responseEntityPlayerData =
@@ -160,15 +160,23 @@ public class PlayerController {
                 restTemplate.getForObject(URL_PROTOCOL + typeTamagotchiServiceBaseUrl + "/types/{typeName}",
                         TypeTamagotchi.class, typeName);
 
-        assert retrievedPlayerData != null;
-        assert typeTamagotchi != null;
+        // assert retrievedPlayerData != null;
+        // assert typeTamagotchi != null;
         return new Player(retrievedPlayerData, typeTamagotchi);
     }
 
 
     // Allow the player to delete their player and progress made
     @DeleteMapping("/player/{playerDataCode}")
-    public ResponseEntity<Object> deletePlayer(@PathVariable String playerDataCode){
+    public ResponseEntity<Object> deletePlayer(@PathVariable String playerDataCode) throws IOException {
+
+        // Check to validate if the user input is valid
+        if (
+            !playerDataCode.matches("[a-zA-Z0-9]+") // Restrict the playerDataCode to letters and digits only
+        )
+        {
+            throw new IOException();
+        }
 
         restTemplate.delete(URL_PROTOCOL + playerDataServiceBaseUrl + "/playerData/" + playerDataCode);
